PXC-3338: Use wsrep_run_commit_hooks like in mysql-wsrep
PXC had a different wsrep_run_commit_hooks function, and it was also
limited to a DBUG_ASSERT for most of it uses.

This commit cherry-picks the function from the mysql-wsrep source, and
changes the assertions to simple conditions.

An additional flag is also added to the function, because the new
whole-transaction logic isn't compatible with other places where PXC
uses this function. As those calls aren't changed in this commit, the
additional flag restores the old behavior for them.