PXC-2574 : Add more info to the "BF lock wait long" message
Issue
When a BF abort occurs, and the locks are not released, the applier threads
can hang.  When this happens "BF lock wait long" is logged, but with
no information about the transaction that is hanging. So this generic
message doesn't help in figuring out what is being blocked.

Solution
Add more information to help diagnosing/debugging problems.